# docs/swagger.yaml
definitions:
  dtos.AddMessageRequest:
    properties:
      message:
        type: string
    type: object
  dtos.AddMessageResponse:
    properties:
      reply:
        type: string
    type: object
  dtos.ChatMessageResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  dtos.ChatRequest:
    properties:
      message:
        type: string
    type: object
  dtos.ChatThreadResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/dtos.ChatMessageResponse'
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dtos.CreateLogBookEntryRequest:
    properties:
      category:
        type: string
      level:
        type: string
      message:
        type: string
    type: object
...
  dtos.CreateThreadRequest:
    properties:
      title:
        type: string
    type: object
...
  models.LogBookEntry:
    properties:
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      timestamp:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/chat:
    post:
      consumes:
      - application/json
      description: Sends a message to OpenAI with MCP tools; no thread history.
      parameters:
      - description: Chat message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: reply
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat with LLM using MCP tools (single turn)
      tags:
      - Chat
  /api/chat/threads:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ChatThreadResponse'
            type: array
      summary: List chat threads
      tags:
      - Chat
    post:
      consumes:
      - application/json
      parameters:
      - description: Optional title
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateThreadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.ChatThreadResponse'
      summary: Create a chat thread
      tags:
      - Chat
  /api/chat/threads/{id}:
    get:
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ChatThreadResponse'
      summary: Get a chat thread with messages
      tags:
      - Chat
  /api/chat/threads/{id}/messages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AddMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.AddMessageResponse'
      summary: Send a message in a thread and get reply
      tags:
      - Chat
  /api/log_book:
    get:
      description: Get a list of all LogBookEntries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LogBookEntry'
            type: array
      summary: Get a list of LogBookEntries
      tags:
      - LogBookEntry
    post:
      consumes:
      - application/json
      description: Create a new LogBookEntry
      parameters:
      - description: LogBookEntry object
        in: body
        name: log_book
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateLogBookEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LogBookEntry'
      summary: Create a new LogBookEntry
      tags:
      - LogBookEntry
  /api/log_book/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a LogBookEntry by ID
      parameters:
      - description: LogBookEntry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LogBookEntry'
      summary: Delete a LogBookEntry
      tags:
      - LogBookEntry
...
swagger: "2.0"